added condition for deleting an accommodation

+++ src/main/java/com/siit/team24/OpenDoors/service/AccommodationService.java
@@ -1,6 +1,7 @@
 package com.siit.team24.OpenDoors.service;
 
 import com.siit.team24.OpenDoors.dto.accommodation.AccommodationHostDTO;
+import com.siit.team24.OpenDoors.exception.ExistingReservationsException;
 import com.siit.team24.OpenDoors.model.Accommodation;
 import com.siit.team24.OpenDoors.repository.AccommodationRepository;
 import jakarta.persistence.EntityNotFoundException;
@@ -12,8 +13,12 @@ import java.util.Optional;
 
 @Service
 public class AccommodationService {
+
     @Autowired
     private AccommodationRepository accommodationRepository;
+
+    @Autowired
+    private ReservationRequestService reservationRequestService;
 
     public Accommodation findById(Long id) {
         Optional<Accommodation> accommodation = accommodationRepository.findById(id);
@@ -32,15 +37,15 @@ public class AccommodationService {
     }
 
     public void delete(Long id) {
-        //TODO check if there are any reservations confirmed for this accommodation in the future
-        //throw exception if so
+        int reservationsNumber = reservationRequestService.countConfirmedFutureFor(id);
+        if (reservationsNumber != 0) {
+            throw new ExistingReservationsException();
+        }
+        reservationRequestService.denyAllFor(id);
         accommodationRepository.deleteById(id);
     }
 
     public void revive(Long id) {
-//        Accommodation accommodation = findById(id);
-//        accommodation.setDeleted(false);
-//        save(accommodation);
         accommodationRepository.revive(id);
     }
